Change f_count and f_msgcount from short to int, solving the rollover     problem demonstrated by the PR.  MFC to follow.
PR: kern/18346

+++ sys/sys/file.h
@@ -62,8 +62,8 @@ struct file {
 #define	DTYPE_FIFO	4	/* fifo (named pipe) */
 #define	DTYPE_KQUEUE	5	/* event queue */
 	short	f_type;		/* descriptor type */
-	short	f_count;	/* reference count */
-	short	f_msgcount;	/* references from message queue */
+	int	f_count;	/* reference count */
+	int	f_msgcount;	/* references from message queue */
 	struct	ucred *f_cred;	/* credentials associated with descriptor */
 	struct	fileops {
 		int	(*fo_read)	__P((struct file *fp, struct uio *uio,
